Rename `copy.hpp` to `repeat_strings.hpp`
